Move SwapIdPrefixes into Ethereum codec
