+++ packages/iov-ethereum/types/ethereumcodec.d.ts
@@ -38,6 +38,10 @@ export interface EthereumCodecOptions {
      */
     readonly erc20Tokens?: ReadonlyMap<TokenTicker, Erc20Options>;
 }
+export declare enum SwapIdPrefixes {
+    Ether = "ether",
+    Erc20 = "erc20"
+}
 export declare class EthereumCodec implements TxCodec {
     private readonly atomicSwapEtherContractAddress?;
     private readonly erc20Tokens;
